requirements.txt
Django~=3.1
psycopg2-binary>=2.8
djangorestframework~=3.11.1
pytz~=2020.1
randomtimestamp~=1.0.0
django-bootstrap4==2.2.0
geopy==2.0.0
pyproj==3
python-telegram-bot==13.0